svis/requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
openai>=1.12.0
groq>=0.4.0
requests>=2.31.0
